Spinning-Menu-Icon
This Repo Contains two files, One is HTML & the Second is CSS of Awesome Spinning Menu Icon, which we have created here. This menu icon too easy to create and in looks it is also too attractive and less line of code. Code is easy to understand, If you want to watch a Video Tutorial of Spinning Menu Icon, Check out the given link:- https://youtu.be/BGFfK0UHFLo